+++ .gitignore
@@ -1,1 +1,3 @@
 /LFU/nbproject/private/
+/LFU/build/
+/LFU/dist/
