+++ LFU/src/lfu/Pair.java
@@ -1,0 +1,11 @@
+package lfu;
+
+class Pair<U, V> {
+	U first;
+	V second;
+	Pair(){}
+	Pair(U u, V v){
+		first = u;
+		second = v;
+	}
+}
